requirements.txt
anitopy~=2.1.0
requests~=2.28.0
num2words~=0.5.10
Anisearch~=1.0.0
python-devlog~=1.2